+++ src/js/navbar.js
@@ -20,7 +20,7 @@ export default class BasisNavbar {
         this.show(item.children(this.args.submenu));
       });
 
-      item.on('mouseleave', (event) => {
+      item.on('mouseleave blur', (event) => {
         this.hidden(item.children(this.args.submenu));
       });
     });
@@ -28,7 +28,14 @@ export default class BasisNavbar {
     $(this.args.item).each((i, e) => {
       const item = $(e);
       item.on('focusin', (event) => {
-        this.hidden(item.siblings(this.args.item).children(this.args.submenu));
+        this.hidden(item.siblings(this.args.item).find(this.args.submenu));
+      });
+
+      item.find(this.args.subitem).each((i, e) => {
+        const subitem = $(e);
+        subitem.on('focusin', (event) => {
+          this.hidden(subitem.siblings(this.args.subitem).find(this.args.submenu));
+        });
       });
     });
   }